requirements.txt
fastapi==0.104.1
jsonlines==4.0.0
openai==0.27.8
pandas==2.1.3
passlib==1.7.4
pydantic==2.5.2
python_jose
PyYAML==6.0.1
SQLAlchemy==2.0.23
starlette
